made comment its own component so it doesn't rerender the whole thing
